chore: Move refresh button to dataset accordion
