+++ web_ui/src/pages/annotator/components/sidebar/dataset/refresh-button/refresh-button.module.scss
